# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Crowdfunding API Description",
        "title": "Crowdfunding API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/campaigns": {
            "get": {
                "description": "Get All Campaign data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Get All Campaign data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "find by user_id",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/campaign.CampaignFormatter"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Campaign data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Create Campaign data",
                "parameters": [
                    {
                        "description": "Body Required",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/campaign.CreateCampaignInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/campaign.CampaignFormatter"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/campaigns/images": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create Campaign Image data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Save Campaign Image data",
                "parameters": [
                    {
                        "description": "Body Required",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/campaign.CreateCampaignImageInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/helper.UploadImageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/campaigns/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Detail Campaign",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Get Campaign Image data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/campaign.CampaignFormatter"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Campaign",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Campaign"
                ],
                "summary": "Update Campaign data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Campaign Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/campaign.CreateCampaignInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helper.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/campaign.CampaignFormatter"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "campaign.CampaignFormatter": {
            "type": "object",
            "properties": {
                "current_amount": {
                    "type": "integer"
                },
                "goal_amount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "short_description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "campaign.CreateCampaignImageInput": {
            "type": "object",
            "required": [
                "campaign_id"
            ],
            "properties": {
                "campaign_id": {
                    "type": "integer"
                },
                "is_primary": {
                    "type": "boolean"
                }
            }
        },
        "campaign.CreateCampaignInput": {
            "type": "object",
            "required": [
                "description",
                "goal_amount",
                "name",
                "perks",
                "short_description"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "goal_amount": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "perks": {
                    "type": "string"
                },
                "short_description": {
                    "type": "string"
                }
            }
        },
...
        "helper.UploadImageResponse": {
            "type": "object",
            "properties": {
                "is_uploaded": {
                    "type": "boolean"
                }
            }
        },
        "helper.meta": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "This should be a single int representing an HTTP status code",
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/helper.meta"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}